src/alfanous-django/requirements.txt
django
distribute==0.6.31
pyparsing==2.0.1
wsgiref==0.1.2
django-debug-toolbar==1.0
django-debug-toolbar-template-timings==0.5.5
dj-database-url
djangorestframework
dynamic-rest
django-rest-swagger
markdown
django-filter
https://github.com/darklow/django-suit/tarball/v2
django_compressor
csscompressor
brotlipy
django-htmlmin
whitenoise
django-swagger-ui
